requirements.txt
hurry.filesize
pydub
librosa
beautifulsoup4
requests
pytz
pymongo
feedparser
python-dotenv
lxml
